services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
httpx==0.27.2
psycopg[binary]==3.2.1
SQLAlchemy==2.0.34
openai==1.57.0
PyYAML>=6.0
